Color tweaks and author update for Autumn.
A few small changes have been made to the Autumn theme. Global variables should
be displayed in the same color as local variables. Errors use the default color
on a red background (instead of black on white).

I've also updated the list of theme authors so that the Pry theme honors the
Creative Commons license that I've used for Autumn.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>